# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API для работы с Task manager",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список всех задач",
                "operationId": "c641d93887fb50d557f18a7d11d0fd70",
                "responses": {
                    "200": {
                        "description": "Успешно получен список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID задачи",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Название задачи",
                                                "type": "string"
                                            },
                                            "description": {
                                                "description": "Описание задачи",
                                                "type": "string"
                                            },
                                            "status_id": {
                                                "description": "ID статуса задачи",
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "description": "ID пользователя, назначенного на задачу",
                                                "type": "integer"
                                            },
                                            "condition_id": {
                                                "description": "ID состояния задачи",
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "description": "Дата создания задачи",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Дата последнего обновления задачи",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "operationId": "53feb2a5084021acb62923cf3cd7a873",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название задачи",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание задачи",
                                        "type": "string"
                                    },
                                    "status_id": {
                                        "description": "ID статуса задачи",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя, назначенного на задачу",
                                        "type": "integer"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния задачи",
                                        "type": "integer"
                                    },
                                    "projects": {
                                        "description": "Список ID проектов",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID задачи",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Название задачи",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Описание задачи",
                                            "type": "string"
                                        },
                                        "status_id": {
                                            "description": "ID статуса задачи",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID пользователя, назначенного на задачу",
                                            "type": "integer"
                                        },
                                        "condition_id": {
                                            "description": "ID состояния задачи",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Дата создания задачи",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления задачи",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить информацию о задаче",
                "operationId": "23615b21f3984fd9361306821c838ad0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получена информация о задаче",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID задачи",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Название задачи",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Описание задачи",
                                            "type": "string"
                                        },
                                        "status_id": {
                                            "description": "ID статуса задачи",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID пользователя, назначенного на задачу",
                                            "type": "integer"
                                        },
                                        "condition_id": {
                                            "description": "ID состояния задачи",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Дата создания задачи",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления задачи",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "operationId": "18e64b4b16cbee264fa84b775e6b7a91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название задачи",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание задачи",
                                        "type": "string"
                                    },
                                    "status_id": {
                                        "description": "ID статуса задачи",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя, назначенного на задачу",
                                        "type": "integer"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния задачи",
                                        "type": "integer"
                                    },
                                    "projects": {
                                        "description": "Список ID проектов",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID задачи",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Название задачи",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Описание задачи",
                                            "type": "string"
                                        },
                                        "status_id": {
                                            "description": "ID статуса задачи",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID пользователя, назначенного на задачу",
                                            "type": "integer"
                                        },
                                        "condition_id": {
                                            "description": "ID состояния задачи",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Дата создания задачи",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления задачи",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "2c379aab1f7cf16dbf084cee9978de75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}